refactor: split add-repository-dialog into sub-components
Extract RegistryTypeSelector and RegistryConfigForm.
